# src/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "get": {
        "description": "Returns private personal and account details for the authenticated user.",
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User record not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Modifies account profile fields (name, email, birthday, phone, address).",
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user profile information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or invalid date format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Username or email is already taken by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me/avatar": {
      "post": {
        "description": "Uploads and updates the user profile photo. Replaces any existing avatar file on disk.",
        "operationId": "UsersController_updateAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Avatar image file (JPEG, PNG, or WebP; maximum 5MB)"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Avatar successfully uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "File missing or invalid file format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Upload user avatar",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "description": "Removes the avatar file from disk and nullifies the database reference.",
        "operationId": "UsersController_deleteAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Avatar successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMessageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Delete user avatar",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me/likes": {
      "get": {
        "description": "Returns recipes previously liked by the current user, ordered by most recently liked.",
        "operationId": "UsersController_getLikedRecipes",
        "parameters": [
          {
            "name": "lang",
            "required": false,
            "in": "query",
            "description": "Target language for projected recipe titles and descriptions",
            "schema": {
              "enum": [
                "it",
                "en",
                "fr"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recipes liked by the user"
          },
          "401": {
            "description": "Authentication required: missing or invalid accessToken cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get liked recipes",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/search": {
      "get": {
        "description": "Searches users by username substring matching (case-insensitive).",
        "operationId": "UsersController_searchUser",
        "parameters": [
          {
            "name": "value",
            "required": true,
            "in": "query",
            "description": "Username substring search query",
            "schema": {
              "example": "mario",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching usernames",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSearchResultDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Search users by username",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "description": "Returns public profile information (username, avatar, join date) for the specified user.",
        "operationId": "UsersController_getUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Numeric user ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public user profile by ID",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "WeCook API",
    "description": "RESTful backend API specification for the WeCook recipe sharing social platform.\n\n### Authentication\nAuthentication is based on secure, HttpOnly JWT cookies:\n- **`accessToken`**: Short-lived (10 minutes) JWT sent automatically by browsers or HTTP clients on protected endpoints.\n- **`refresh_token`**: Long-lived (7 days) JWT restricted to the `/api/auth/refresh` path.\n\n### Media Storage & Static Assets\nUploaded recipe covers, step images, gallery photos, videos, and user avatars are served statically under `/uploads/`.\n\n### Localization & Multilingual Engine\nRecipes and catalogs support Italian (`it`), English (`en`), and French (`fr`), powered by LibreTranslate relational projection.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://app.we-cook.it:8443",
      "description": "Staging / Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token",
        "description": "JWT Refresh Token stored in HttpOnly cookie (`refresh_token`) for the refresh endpoint"
      }
    },
    "schemas": {
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code representing the error category",
            "example": 400
          },
          "timestamp": {
            "type": "string",
            "description": "ISO 8601 formatted timestamp of when the error occurred",
            "example": "2026-10-09T17:30:00.000Z"
          },
          "path": {
            "type": "string",
            "description": "Requested URL path that produced the error",
            "example": "/api/recipes/42"
          },
          "message": {
            "description": "Error description or list of validation error messages",
            "example": "Recipe not found",
            "oneOf": [
              {
                "type": "string",
                "example": "Recipe not found"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "Title must be at least 3 characters long"
                ]
              }
            ]
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "message"
        ]
      },
...
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User identifier",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Unique username",
            "example": "mario_rossi"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "mario.rossi@example.com"
          },
          "role": {
            "description": "User role",
            "example": "user",
            "enum": [
              "admin",
              "moderator",
              "user"
            ],
            "type": "string"
          },
          "avatar_url": {
            "type": "string",
            "nullable": true,
            "description": "Avatar image URL",
            "example": "/uploads/avatars/avatar_1.webp"
          },
          "first_name": {
            "type": "string",
            "nullable": true,
            "description": "First name",
            "example": "Mario"
          },
          "last_name": {
            "type": "string",
            "nullable": true,
            "description": "Last name",
            "example": "Rossi"
          },
          "birth_date": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Date of birth",
            "example": "1995-05-20T00:00:00.000Z"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "description": "Phone number",
            "example": "+39 06 1234567"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "description": "Residential street address",
            "example": "Via Roma 123"
          },
          "city": {
            "type": "string",
            "nullable": true,
            "description": "City of residence",
            "example": "Roma"
          },
          "postal_code": {
            "type": "string",
            "nullable": true,
            "description": "Postal code / ZIP",
            "example": "00100"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Account creation timestamp",
            "example": "2026-10-09T18:00:00.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role",
          "avatar_url",
          "first_name",
          "last_name",
          "birth_date",
          "phone",
          "address",
          "city",
          "postal_code",
          "created_at"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Updated username (3-30 characters)",
            "example": "mario_rossi_updated",
            "minLength": 3,
            "maxLength": 30
          },
          "email": {
            "type": "string",
            "description": "Updated email address",
            "example": "mario.updated@example.com",
            "format": "email"
          },
          "first_name": {
            "type": "string",
            "description": "First name",
            "example": "Mario",
            "maxLength": 100
          },
          "last_name": {
            "type": "string",
            "description": "Last name",
            "example": "Rossi",
            "maxLength": 100
          },
          "birth_date": {
            "type": "object",
            "description": "Date of birth (ISO 8601 date string or timestamp)",
            "example": "1995-05-20"
          },
          "phone": {
            "type": "string",
            "description": "Contact phone number",
            "example": "+39 06 1234567",
            "maxLength": 30
          },
          "address": {
            "type": "string",
            "description": "Residential street address",
            "example": "Via Roma 123",
            "maxLength": 255
          },
          "city": {
            "type": "string",
            "description": "City of residence",
            "example": "Roma",
            "maxLength": 100
          },
          "postal_code": {
            "type": "string",
            "description": "Postal code / ZIP",
            "example": "00100",
            "maxLength": 10
          }
        }
      },
      "AvatarUploadResponseDto": {
        "type": "object",
        "properties": {
          "avatar_url": {
            "type": "string",
            "description": "Static URL path to newly saved avatar photo",
            "example": "/uploads/avatars/avatar_1.webp"
          }
        },
        "required": [
          "avatar_url"
        ]
      },
      "UserMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Status message",
            "example": "Avatar deleted successfully"
          }
        },
        "required": [
          "message"
        ]
      },
      "UserSearchResultDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Matching username",
            "example": "mario_rossi"
          }
        },
        "required": [
          "username"
        ]
      },
      "PublicUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User identifier",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "mario_rossi"
          },
          "avatar_url": {
            "type": "string",
            "nullable": true,
            "description": "Public avatar photo URL",
            "example": "/uploads/avatars/avatar_1.webp"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Member since date",
            "example": "2026-10-09T18:00:00.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "avatar_url",
          "created_at"
        ]
      },
...
    }
  }
}